Finished and bug tested initial build of pi server

+++ build/RasPi/python_programs/data_processor.py
@@ -1,31 +1,31 @@
 import sys
 import json
+import os.path
 
-class HandleInput:
+def writeJSONCSV(inputStr,CSVFile):
 
-    @staticmethod
-    def writeJSONCSV(inputStr,CSVFile):
+    jsonData = json.loads(inputStr)
+    CSVOutput = JSONtoCSV(jsonData)
 
-        jsonData = json.loads(inputStr)
-
-        CSVOutput = JSONtoCSV(jsonData)
-
-        with open(CSVFile, 'w') as f:
-            try:
+    if os.path.isfile(CSVFile):
+        try:
+            with open(CSVFile, "a") as f:
                 f.write(CSVOutput)
                 f.close()
-            except:
-                try:
-                    print("CSV file does not exist\nCreating new file")
-                    with open(CSVFile, "w+") as fn:
-                        fn.write("time,device,sensor,value")
-                        fn.write(CSVOutput)
-                        fn.close()
-                except:
-                    print("CSV file does not exist\n")
-                    exit()
+        except Exception as e:
+            print("\nError printing to \n")
+            exit() #add exception here to keep things running maybe add error log
+    else:
+        try:
+            print("CSV file does not exist\nCreating new file")
+            with open(CSVFile, "w+") as fn:
+                fn.write("time,device,sensor,value\n")
+                fn.write(CSVOutput + "\n")
+                fn.close()
+        except:
+            print("Tried to create file. CSV file does not exist\n")
+            exit()
 
-    @staticmethod
-    def JSONtoCSV(jsonData):
-        csvOut = jsonData["time"] + ", " + jsonData["device"] + ", " +  jsonData["sensor"] + ", " + jsonData["value"]
-        return csvOut
+def JSONtoCSV(jsonData):
+    csvOut = jsonData["time"] + ", " + jsonData["device"] + ", " +  jsonData["sensor"] + ", " + str(jsonData["value"])
+    return csvOut
